requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
jinja2>=3.1.2
aiohttp>=3.8.5
apscheduler>=3.10.1
openai>=1.0.0
